+++ package.json
@@ -20,6 +20,7 @@
         "react-hooks-helper": "^1.6.0",
         "react-iframe": "^1.8.0",
         "react-scripts": "4.0.1",
+        "react-select": "^5.2.1",
         "styled-components": "^5.3.3",
         "typescript": "^4.5.2",
         "web-vitals": "^0.2.4"
