+++ src/services/index.tsx
@@ -1,0 +1,27 @@
+import { initializeApp } from 'firebase/app';
+
+import { getFirestore, getDocs, collection } from 'firebase/firestore/lite';
+import { getAnalytics } from "firebase/analytics";
+
+const firebaseConfig = {
+  apiKey: "AIzaSyDv-nyeH2lmdGy0fU9FJTrCaaf4wMnZD9g",
+  authDomain: "grio-53e46.firebaseapp.com",
+  projectId: "grio-53e46",
+  storageBucket: "grio-53e46.appspot.com",
+  messagingSenderId: "141665876430",
+  appId: "1:141665876430:web:cffe0876202547ebdf539f",
+  measurementId: "G-Y3M71N3PHX"
+};
+
+const app = initializeApp(firebaseConfig);
+const analytics = getAnalytics(app);
+
+
+
+export async function getAreas() {
+  const db = getFirestore(app);
+  const areasCol = collection(db, 'areas');
+  const areaSnapshot = await getDocs(areasCol);
+  const areasList = areaSnapshot.docs.map(doc => doc.data());
+  return areasList;
+}
